return float

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -2,7 +2,7 @@
 -- Safely and returns 0 if the quotient is 0
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS DECIMAL
+RETURNS FLOAT
 BEGIN
   IF b = 0 THEN
     RETURN 0;
